move firebase.ts to lib
