+++ src/middleware/authenticated.ts
@@ -1,10 +1,10 @@
 import { Context } from '@nuxt/vue-app'
-import { auth } from '@/plugins/firebase'
+import { auth } from '@/lib/firebase'
 
 export default (context: Context) => {
   auth.onAuthStateChanged((user) => {
     if (user) {
       context.store.dispatch('user/signin', user)
-    } else if (context.route.name !== 'signin') context.redirect('/signin')
+    }
   })
 }
